% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Create Pico Package Skeleton}
\usage{
create(name, dir)
}
\arguments{
\item{name}{Character. Name for the package.}

\item{dir}{Character. Folder location to write the package.}
}
\value{
Nothing. Directory written or not.
}
\description{
Generate a directory containing the absolute bare minimum content
for a package.
}
\examples{
\dontrun{create("mypkg", "~/Desktop")}
}
